Move memory status function to helper lvl
